Deposit Reward 0 Gold Patch

+++ ZipModScript.bat
@@ -2,10 +2,10 @@
 setlocal
 
 :: Define directories and filenames
-set "scriptDirInterkarmaWindows=D:\Downloads\daggerfall-unity-1.1.1\StandaloneWindows"
-set "scriptDirInterkarmaLinux=D:\Downloads\daggerfall-unity-1.1.1\StandaloneLinux64"
-set "scriptDirInterkarmaOsx=D:\Downloads\daggerfall-unity-1.1.1\StandaloneOSX"
-set "scriptDirVwing=D:\Downloads\daggerfall-unity-android-android\Android"
+set "scriptDirInterkarmaWindows=C:\Users\Admin\Downloads\daggerfall-unity-1.1.1\daggerfall-unity-1.1.1\StandaloneWindows"
+set "scriptDirInterkarmaLinux=C:\Users\Admin\Downloads\daggerfall-unity-1.1.1\daggerfall-unity-1.1.1\StandaloneLinux64"
+set "scriptDirInterkarmaOsx=C:\Users\Admin\Downloads\daggerfall-unity-1.1.1\daggerfall-unity-1.1.1\StandaloneOSX"
+set "scriptDirVwing=C:\Users\Admin\Downloads\daggerfall-unity-android-android\daggerfall-unity-android-android\Android"
 
 set "fileName=banks remaked.dfmod"
 set "zipFileNameWindows=banks_remakedWIN.zip"
@@ -29,25 +29,25 @@ set "scriptDir=%~1"
 set "fileName=%~2"
 set "zipFileName=%~3"
 
-:: Set file and zip paths
+:: Set paths for file, Mods folder, and zip file
 set "filePath=%scriptDir%\%fileName%"
+set "modsDir=%scriptDir%\Mods"
 set "zipFilePath=%scriptDir%\%zipFileName%"
-set "tempDir=%scriptDir%\temp"
 
-:: Create temporary directory if it doesn't exist
-if not exist "%tempDir%" mkdir "%tempDir%"
+:: Create Mods directory if it doesn't exist
+if not exist "%modsDir%" mkdir "%modsDir%"
 
-:: Copy the file to the temporary directory
-copy "%filePath%" "%tempDir%\%fileName%"
+:: Move the file to the Mods directory
+move /y "%filePath%" "%modsDir%\%fileName%"
 
 :: Delete existing zip file
-del /q "%zipFilePath%"
+if exist "%zipFilePath%" del /q "%zipFilePath%"
 
-:: Create the zip file
-powershell -command "Add-Type -A 'System.IO.Compression.FileSystem'; [IO.Compression.ZipFile]::CreateFromDirectory('%tempDir%', '%zipFilePath%')"
+:: Create the zip file from the Mods directory
+powershell -command "Add-Type -A 'System.IO.Compression.FileSystem'; [IO.Compression.ZipFile]::CreateFromDirectory('%modsDir%', '%zipFilePath%')"
 
-:: Clean up temporary directory
-rd /s /q "%tempDir%"
+:: Clean up Mods directory (if desired, comment this out if you want to keep the Mods directory)
+rd /s /q "%modsDir%"
 
 endlocal
 exit /b
